+++ example/android/app/src/main/java/com/example/reactnativepanortc/MainApplication.java
@@ -2,6 +2,9 @@ package com.example.reactnativepanortc;
 
 import android.app.Application;
 import android.content.Context;
+
+import androidx.multidex.MultiDex;
+
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
@@ -41,6 +44,13 @@ public class MainApplication extends Application implements ReactApplication {
   public ReactNativeHost getReactNativeHost() {
     return mReactNativeHost;
   }
+
+    @Override
+    protected void attachBaseContext(Context base) {
+        super.attachBaseContext(base);
+
+        MultiDex.install(this);
+    }
 
   @Override
   public void onCreate() {